+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/forwarders/PhotoForwarder.kt
@@ -34,11 +34,11 @@ class PhotoForwarder : Forwarder {
                 postMessage to it
             }
         }.map {
-            pair ->
-            bot.executeSync(pair.second).let {
+            (original, request) ->
+            bot.executeSync(request).let {
                 response ->
                 response.message() ?.let {
-                    pair.first to it
+                    original to it
                 } ?:let {
                     throw IOException("${response.errorCode()}: ${response.description()}")
                 }