+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/forwarders/TextForwarder.kt
@@ -27,11 +27,11 @@ class TextForwarder : Forwarder {
                 ParseMode.Markdown
             )
         }.map {
-            pair ->
-            bot.executeSync(pair.second).let {
+            (original, request) ->
+            bot.executeSync(request).let {
                 response ->
                 response.message() ?.let {
-                    pair.first to it
+                    original to it
                 } ?:let {
                     throw IOException("${response.errorCode()}: ${response.description()}")
                 }