+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/forwarders/VoiceForwarder.kt
@@ -33,11 +33,11 @@ class VoiceForwarder : Forwarder {
                 }
             }
         }.map {
-            pair->
-            bot.executeSync(pair.second).let {
+            (original, request) ->
+            bot.executeSync(request).let {
                 response ->
                 response.message() ?.let {
-                    pair.first to it
+                    original to it
                 } ?:let {
                     throw IOException("${response.errorCode()}: ${response.description()}")
                 }