+++ CHANGELOG.md
@@ -72,3 +72,12 @@ which block all requests data thread)
 * Added `/getPublishSchedule` command
 * Added `/disableSchedulePublish` command
 * Now publisher will automatically remove `/publishPost` message
+
+## 0.4.3
+
+* **Now posts contain creation datetime** (look at `PostsTable#getPostCreationDateTime(Int)`)
+* Update `BotLogger` behaviour
+* Fix problem with canceling of next scheduled job for `SchedulerPlugin`
+* Fix behaviour of GetSchedulesCommand when queue is empty
+* Add for items of `SmartChooser` plugin config options `minAge` and `maxAge` for prefilter
+posts which too old or too young